factor out mapping logic
