+++ api/platform_hub/services.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 from collections import defaultdict
 from datetime import timedelta
-from typing import Any
 
 import structlog
 from django.conf import settings
@@ -20,6 +19,7 @@ from organisations.models import (
     OrganisationSubscriptionInformationCache,
 )
 from platform_hub.constants import get_integration_registry
+from platform_hub.mappers import map_release_pipeline_stage_to_stats_data
 from platform_hub.types import (
     EnvironmentMetricsData,
     IntegrationBreakdownData,
@@ -579,25 +579,9 @@ def get_release_pipeline_stats(
         )
         total_features += completed_features
 
-        stages_data: list[ReleasePipelineStageStatsData] = []
-        for stage in stages_qs:
-            features_in_stage = stage.get_in_stage_feature_versions_qs().count()
-            features_completed = stage.get_completed_feature_versions_qs().count()
-
-            action_description = _build_action_description(stage)
-            trigger_description = _build_trigger_description(stage)
-
-            stages_data.append(
-                ReleasePipelineStageStatsData(
-                    stage_name=stage.name,
-                    environment_name=stage.environment.name,
-                    order=stage.order,
-                    features_in_stage=features_in_stage,
-                    features_completed=features_completed,
-                    action_description=action_description,
-                    trigger_description=trigger_description,
-                )
-            )
+        stages_data: list[ReleasePipelineStageStatsData] = [
+            map_release_pipeline_stage_to_stats_data(stage) for stage in stages_qs
+        ]
 
         results.append(
             ReleasePipelineOverviewData(
@@ -615,27 +599,3 @@ def get_release_pipeline_stats(
         )
 
     return results
-
-
-def _build_action_description(stage: Any) -> str:
-    """Build a human-readable description of the stage's actions."""
-    actions = stage.actions.all()
-    if not actions:
-        return ""
-    descriptions: list[str] = []
-    for action in actions:
-        descriptions.append(action.get_action_type_display())
-    return ", ".join(descriptions)
-
-
-def _build_trigger_description(stage: Any) -> str:
-    """Build a human-readable description of the stage's trigger."""
-    try:
-        trigger = stage.trigger
-    except stage.__class__.trigger.RelatedObjectDoesNotExist:
-        return ""
-
-    description: str = trigger.get_trigger_type_display()
-    if trigger.trigger_body and "wait_for" in trigger.trigger_body:
-        description += f" ({trigger.trigger_body['wait_for']})"
-    return description
